Make MyPermissivePolicy even more permissive
I spent a couple hours trying to debug errors resulting from me thinking that this policy is a good base wildcarded policy for testing device auth. My client devices name didn't start with "MyClientDevice" so I was getting weird MQTT protocol errors on the client side.

I think that because of the examples above, you're still getting to show the example of what if we wanted to match only client devices that match "MyClientDevice*" but you would also get an extremely permissive example base policy to do test development on, and for a policy that is called permissive, I think it would be less confusing if the whole policy document was extremely permissive.